Rename load testing service to flagger-loadtester
